+++ src/sanity/schema/annotations/index.ts
@@ -1,0 +1,2 @@
+export * from "./externalLink";
+export * from "./internalLink";
