+++ src/sanity/schema/blocks/index.ts
@@ -1,0 +1,5 @@
+export * from "./blockquote";
+export * from "./break";
+export * from "./image";
+export * from "./richText";
+export * from "./youtube";
